Solve 'the bug'!
"The 128-byte area beyond the location pointed to by %rsp is considered
to be reserved and shall not be modified by signal or interrupt
handlers. Therefore, functions may use this area for temporary data that
is not needed across function calls. In particular, leaf functions may
use this area for their entire stack frame, rather than adjusting the
stack pointer in the prologue and epilogue. This area is known as the
red zone."

We don't support red-zone just yet haha.