pkg: move zasm.zig to boot in preparation for Zig 0.11.0 update
In Zig 0.11.0, the package management system was rewritten,
and as such the support for --begin-pkg was removed.

To simplify the story of package management now, simply include
zasm.zig as a file (rather than placing it in a separate Git
repository).

Since Zig doesn't allow local import of files outside of the
root directory of the Zig file that is using the import, move
zasm.zig to the boot/ directory.