gdt64: working userland with hard-coded offset
Problem with nasm compile-time and relative section offset calculations
is that they are very brittle. The old value was probably true in some
previous instance of this code. However, the same is true for hard-coded
offsets as well, but since the next step is to move the gdt related code
into a higher level language, we can keep this commit as a integration
test while migrating.

`0x10b0ee` is the location of the tss64 struct in memory. When debugging
it is mostly empty with only +0x4 set to a pointer value of the kernel
stack (.rsp0).