paging: nice virtual ASCII memory mappings.
Refactor of paging_32.zig now allows for virtual addresses of up-to 4GB.
Now implementing a state-of-the-art ASCII compatible virtual address
scheme for next-gen debugging performance.

Specifically map the "bootstrap" page to start the kernel. In time we'll
probably zero out this segment and page-out the null page.

Move the non-temporary stack position calculation to zig.