run/debug: Increase memory demands from 128M -> 4G
Since we are using a very naive virtual paging algorithm which basically
maps virtual to physical 1-to-1, we use a _lot_ of memory.

A clear improvement would be to track how many pages we actually use
instead of calculating the base address from the page index. That way we
can reduce the memory footprint to probably less than 1G.